+++ src/neon/alpha_f16_full.rs
@@ -31,12 +31,13 @@ use std::arch::aarch64::*;
 
 use crate::alpha_handle_f16::{premultiply_pixel_f16_row, unpremultiply_pixel_f16_row};
 use crate::neon::f16_utils::*;
+use core::f16;
 use rayon::iter::{IndexedParallelIterator, ParallelIterator};
 use rayon::prelude::{ParallelSlice, ParallelSliceMut};
 use rayon::ThreadPool;
 
 #[target_feature(enable = "fp16")]
-unsafe fn neon_premultiply_alpha_rgba_row_f16_full(dst: &mut [half::f16], src: &[half::f16]) {
+unsafe fn neon_premultiply_alpha_rgba_row_f16_full(dst: &mut [f16], src: &[f16]) {
     let mut rem = dst;
     let mut src_rem = src;
 
@@ -66,9 +67,9 @@ unsafe fn neon_premultiply_alpha_rgba_row_f16_full(dst: &mut [half::f16], src: &
 }
 
 pub(crate) fn neon_premultiply_alpha_rgba_f16_full(
-    dst: &mut [half::f16],
+    dst: &mut [f16],
     dst_stride: usize,
-    src: &[half::f16],
+    src: &[f16],
     src_stride: usize,
     width: usize,
     _: usize,
@@ -95,7 +96,7 @@ pub(crate) fn neon_premultiply_alpha_rgba_f16_full(
 }
 
 #[target_feature(enable = "fp16")]
-unsafe fn neon_unpremultiply_alpha_rgba_f16_row_full(in_place: &mut [half::f16]) {
+unsafe fn neon_unpremultiply_alpha_rgba_f16_row_full(in_place: &mut [f16]) {
     let mut rem = in_place;
 
     for dst in rem.chunks_exact_mut(8 * 4) {
@@ -137,7 +138,7 @@ unsafe fn neon_unpremultiply_alpha_rgba_f16_row_full(in_place: &mut [half::f16])
 }
 
 pub(crate) fn neon_unpremultiply_alpha_rgba_f16_full(
-    in_place: &mut [half::f16],
+    in_place: &mut [f16],
     stride: usize,
     width: usize,
     _: usize,
